% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Display country population}
\usage{
CountryPopulation(x)
}
\arguments{
\item{x}{A country name as a string}
}
\value{
The graph of a countries' population over time as \code{y}
}
\description{
Function to display population over time for a specified countryme
Fizz Buzz.
}
\examples{
CountryPopulation('United States of America')
CountryPopulation('Russian Federation')
}
