% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Attendance and matches of all World Cups}
\format{
A data frame of 22 observations
\describe{
   \item{TotalAttendance}{Total attendance for that particular world cup}
   \item{Matches}{Number of matches for the world cup}
   \item{AverageAttendance}{Total attendance for that particular world cup}
   \item{WorldCup}{The year and host country of the world cup}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
Attendance and matches of all World Cups
}
\keyword{datasets}
