% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Population.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Populations of Countries from 1950-2020}
\format{
A data frame of 16685 observations
\describe{
   \item{Country}{The name of the Country}
   \item{Year}{Year of observation}
   \item{Population}{Population for the Year}
}
}
\source{
\file{World_Population.xlsx}
}
\usage{
WorldPopulation
}
\description{
Populations of Countries from 1950-2020
}
\keyword{datasets}
